resolution: official names for Ceilometer and Heat
During the Board of Directors meeting on Monday, it became clear that
there doesn't appear to be any controversy around allowing Ceilometer
and Heat use their official names.

The formal process in the bylaws is that the TC must recommend to the
board that they include the projects in the "Core OpenStack Project".

This commit proposes a resolution to this effect. The resolution also
makes it clear that we are not asking the board to require the use of
Heat and Ceilometer in products using the OpenStack trademark.

Change-Id: Ic751aa28b8f88e8c8d7f15b76bc1825725ac4c05