Incubation / Graduation / New program requirements
The TC applies a number of criteria when considering project incubation
and graduation to integrated status, and when considering new programs.
These reference documents should track the minimal requirements that we
apply, so that candidate projects and teams know what is expected of
them.

These documents, by nature, do not contain all the criteria that each
member of the TC will apply when evaluating applications. They are
just the ever-evolving minimal set of requirements that are consensual
amongst TC members.

NB: Some currently-incubated, currently-integrated or currently-official
projects and programs do not fill all of those requirements, as some of
them are new requirements we recently imposed to raise the quality of
the end-result, based on past experience. Those teams and projects
should try to cover the gap during the Icehouse cycle.

Change-Id: Ieedc64a20135e3498786bee6fc54d91b9cf2a93a