provide infrastructure for publishing docs to html
this brings over the pbr infrastructure for publishing rst files
to html with the oslo.sphix style. We'll need to move around
and reformat some of the documents for them to fully take effect
but this would be the first step in getting there.

Updated with our latest knowledge on how to do the right tox.ini
changes to handle sphinx issues.

This will then get used by another publish job to go to some
online repository.

Change-Id: I0dd3ecf0a9b3e356a4a2980a305b421b7d7b8602