Attach projects to programs
List projects (repositories) attached under each program. Also tracks
if a given project in a program is integrated or incubated.

PTLs oversee projects under their program. Contributing to a project
attached to a program gives you the right to participate in that program
PTL election.

All openstack*/* projects are assigned to a program, except:
- openstack/openstack, openstack/governance and openstack/requirements
  which are affecting everyone (so under the TC supervision)
- openstack-dev/openstack-qa, openstack/melange,
  openstack/openstack-chef, and openstack/python-melangeclient which
  should move to some attic
- openstack-dev/sandbox and openstack/python-openstackclient which
  should be in stackforge

Change-Id: Iac3b71f56e138b582f4f8d1a1abfa1068ac46c55