Add a requirement for deprecating duplicated code
This patch updates graduation requirements related to scope.  The requirements
currently say a project must not duplicate the scope of another project.
However, we have some cases where a project intentionally duplicates
functionality with the intent of replacing it (Neutron, Cinder).

In that case, I believe we should have an additional requirement that the new
project may not graduate until it has reached a level of functionality and
maturity where we're able to deprecate the component it is replacing with a
known timeline for removal.

Change-Id: I5305e2f80ebc3c78ef74b0957337fe11c3f3de08