Remove gantt from Compute Program
At the Midcycle nova meetup we decided that the Compute Program is not
ready to host the Gantt repository yet. This is because we don't want to
put the review burdon on nova-core while at the same time opening up the
repo to a whole other group of core reviewers.

http://lists.openstack.org/pipermail/openstack-dev/2014-February/027370.html

Change-Id: Ib12a0f2b571e6ab33a7d3a6bef36ee56b16edad4