Add the requirements repo to the release program
At the project meeting on 2014-03-18 [1] we discussed some changes to
the way global requirements are managed, especially to ensure that
the team reviewing changes in that repository has a leader and that
the review criteria are tightened up. Thierry was nominated as PTL, as
the current Release Manager. The repository is not currently managed by
any program, and it makes more sense to me to put it under the existing
release program rather than creating a new one.

[1] http://eavesdrop.openstack.org/meetings/project/2014/project.2014-03-18-21.03.log.txt

Change-Id: I3369db7bfe74a129d9c2d4dad384a360ded3aa02