Add the Kite key distribution service to programs.yaml
The Kite project has been moved under the Key Management Service program.
Hence this CR adds the Kite repository to the Key Management Service section
of the programs.yaml file.

The Kite project is an implementation of a Key Distribution Service, which
is designed to enable secure messaging.  Some background is available at the
following locations:

  https://wiki.openstack.org/wiki/MessageSecurity
  https://github.com/cloudkeep/barbican/wiki/Blueprint:-KDS-Service

A recent thread about the decision to place Kite under the Key Management
Service program is available here:

  http://lists.openstack.org/pipermail/openstack-dev/2013-November/thread.html#19998

Change-Id: I62c01c398e02de98f85267d60efc8bfbbf1d119b