add upgrade testing expectation
This was part of the original QA proposal for project requirements,
however when we lost the post integration section it was lost. We
should bring back the expectation of upgrade testing, as one of
the top community concerns by operators it the ability to upgrade
their clouds from one release to the next.

Change-Id: Ie01d2a8b27e7cf9a0ec924c1d11d748af364929e