begin conversion to rst
This converts 2 of the reference documents to rst, and fixes the
tox.ini so that tox -e docs creates a pretty html version of these
documents. While the publish target location has still not been
sorted this moves the ball forward in generating nicer output
for the documents.

The incubation-integration-requirements document was intentionally
left out, because it's much more active, and we should decide this
is the right approach before we impose the merge conflicts on
everyone in flight.

Change-Id: I2fefdb5efe196cdbda4555141b09e7e771e20bdd