add upgrade expectations
While we referenced upgrade in the QA expectations, we didn't yet
have any language on upgrade. Because the CDing upgrade expectation
often comes as a surprise to projects, we should be explicit about
it.

This is intentionally a more "folksy" version of the language, as the
more legalistic version appeared to be generating confusion.

Change-Id: Ic73d490d0ed4841600a00e3433163c3bd93a2660