2014-07-11 Election Activities
This draft represents a combination of the work done in two earlier
versions of the resolution ([anteaya][eglynn]) and notes from the
TC meeting on 8 July 2014 ([logs][markmc]).

This version adds details based on the background for each position,
so that readers who have not followed the discussion will have a
better understanding of why we are talking about these issues. It
also tries to provide a clear reporting structure and cover some
of the technicalities of who can administer the election while some
members of the TC are running.

It also focuses on resolving issues amicably and educating the
community, rather than publicizing faults in a way that may punish
someone harshly for a simple mistake. I think this is in keeping
with the "benefit of the doubt" attitude I sensed from the most
recent meeting.

[anteaya] https://review.openstack.org/#/c/98675
[eglynn] https://review.openstack.org/#/c/100445
[logs] http://eavesdrop.openstack.org/meetings/tc/2014/tc.2014-07-08-20.02.log.html
[markmc] http://paste.openstack.org/raw/85709/

Change-Id: I2c3cab078b16311c174a122960f223577d6577bf