Expand scope of the Identity program to include auditing
Authentication and authorization is always closely followed by auditing
how the user obtained that authorization and what the user does with
that authorization.

As OpenStack's approach to auditing has matured, it has become clear
that that there is a regular cross-pollination of contributors and
stakeholders between the two projects, making it appropriate to
officially integrate auditing into the Identity program's mission.

As a result, this change also moves the PyCADF library to the Identity
program, which provides OpenStack's auditing data model.

Change-Id: I0cd840e88b6f5ec9e576e39f4dff739cba3f04c1