Add Juno Compute co-authored-by authors to extra-atcs.
As discussed on our mailing list, I'd like to treat juno co-authors
for Compute ATCs. This is the simplest way to achieve that.

I have been asked to verify the foundation membership status of the
various people proposed in this patch. There is a comment per
person which does that. If the person has a commit _ever_ in an
incubated or graduated repository, then I consider them to be a
foundation member. For those authors where that is not true I have
hand verified their foundation profile id.

I have also been asked to _not_ generate this for other programs,
so I have stopped doing that and gone back to just doing Compute.
Other PTLs can do this for their programs if they desire, using
the script I will propose in a separate change.

Change-Id: I687b59e037593499adcb112e2bf348afc5f2fb7d