Import the Project Testing Interface description
This text was approved by the PPB way back in OpenStack's history. There
are several things which have changed and have not been recorded in it,
and there is a new change that should be made. For now, this is verbatim
what was voted on in the past. Changes will follow to update to current
reality.

Change-Id: I9b5e0e5a52aab1d00497d2e8a75e9c9b249a27e9