Recommendation to Adopt DCO as CLA
At its July 2014 meeting, the OpenStack Foundation Board of Directors
heard information about the impact and options to address the
OpenStack Contributor License Agreement.  One of the questions raised
was whether the issues being brought forward were simply the opinion
of one member of the Board.

This resolution clarifies that the TC believes that the current CLA is
problematic and the DCO should be considered instead.

Change-Id: I5da90bd9ae3c7c08c028aa8dd3d0521b8b44d593