Add a new openstack-specs repo under the control of the TC
One of the takeaways from the exercise about kilo priorities was oslo is
not the right home for all cross project discussions. We need a new home
for cross project specs, and since the TC is our governing technical
committee this repo should be under there management.

The specific scope and model for this repo is left up to the TC to
decide. Although I am in favor of using this for specs that are not limited
to a just a few projects (although not every project may hit the issue
the spec is addressing at the same time) and using this as recommended
guidelines. For example having a spec on doing API microversioning would
be useful as several projects hope to implement this and it would be
ashame if they all did it slightly differently. Another example is
logging standards: what debug,info etc should consist of.

Related project-config change: Id9393dfc2c7f1f883a5f20d85f7ecb79e2c52063

Change-Id: I498a304057043a54370b132f2153a4c65840639c