Add heat-translator to the Heat program
After having a good discussion on the IRC Heat weekly meeting and mailing
list, the Heat PTL and core team voted and agreed to have the Heat-Translator
project added to its official program. Heat-Translator is currently a
StackForge project and provides a way to translate non-Heat templates to
Heat Orchestration Template (HOT).

Change-Id: I038b9e9ac6fc2b3584dcbc5ffd1033fa8fb0d52e