Add Castellan project to Barbican program
After reviewing a few projects that are starting to integrate with
Barbican we noticed that a couple of them were defining a Key Manager
Interface and then implementing the interface using
python-barbicanclient instead of integrating with Barbican directly.

We were able to identify a couple of advantages with this pattern.
Namely that it gives the projects flexibilty to choose an alternative
Key Manager as needed.  This is helpful during development, for example,
where a simple implementation of the Key Manager Interface can be used
instead of having to spin up a Barbican instance.  It would also be
helpful in deployments with strict compliance requirements where a Key
Manager with specific certifications is required which would prevent the
usage of Barbican.

In order to prevent duplication of efforts, the Barbican team would like
to provide the generic Key Manager interface, as well as the
python-barbicanclient implementation of the interface.

We asked the Oslo team about creating a new oslo library for this
purpose, but Doug Hellman's recommendation was that we create a new
repository under the Barbican program since he would prefer that the
Barbican team be responsible for reviewing changes to the Key Manager
interface.

Following Doug Hellman's recommendation, and after many hours of
bikeshedding a name for the new repository, the Barbican team agreed on
creating a new repository called "Castellan" to hold the generic Key
Manager Interface.

We have requested a new repository in the following CR:

https://review.openstack.org/#/c/136140/

As requested by the Infra team, we're also submitting this CR to the
governance repo for review by the TC.

Change-Id: I0e1909e767da1cf696df126450caa6261657c77d