Adding the Murano Application Catalog to OpenStack
Murano is a set of code repositories that implements an application catalog.
Our mission is to provide a RESTful API and user interface that allows users
to compose and deploy complex composite application environments and manage
the lifecycle of those environments. The code base has existed for well
over 2 years in the stackforge code namespace.

The Murano contributor community being made up of more
than 10 contributors from Mirantis, HP, Telefonica, and
other organizations, and we hope to get it more diverse
after moving to /openstack space.
We will hold elections for our PTL, currently Serg Melikyan, on a regular basis
according to standard OpenStack community practices. Our meetings are weekly,
on IRC, and our roadmap is transparent and managed on Launchpad again
using standard OpenStack community practices.

We rely on Heat as orchestration engine for provisioning application on vm,
and plan to move storing application packages to Glance, once
Artifact Repository will be available in Glance.

Change-Id: I24b5309a2365495e041155ee2d6799dbf79479d9