Adding the security team.
The VMT and the OSSG have merged to form a single security team.

We are seeking to have this team recognised as a horizontal project
within OpenStack, similar in many ways to the current documentation
team.

Advisories, Documentation, Security Notes, Threat Analysis etc
are all measured through their respective respositories.

Change-Id: I9d0db560952380d3283ca710608ed0fdd3af64da