Rename keystoneclient-federation to saml2
Federation is becoming a very core part of keystone's scope. We had
initially split keystoneclient-federation particularly to separate
additional dependencies. It was decided that it is more correct to call
this repository keystoneclient-saml2 as this is what will be there and
we may not want to lump together all federation plugins in one repo.

This library has never had a release so will not have an impact on
users.

Change-Id: Ia25e14a6afa939590ebba3a91973ceee4a8eaee2