Add tool/framework to automatically validate the tag applications.
Add tools/validate_tags.py framework to automatically validate the usage
of tags in reference/projects.yaml. This is only for tags that can be
enforced pragmatically.

This is not intended for gating purposes, rather to make the process of
updating projects.yaml easier.

Supports both team and repository based tags.

Updated diversity.py to work with validate_tags

Change-Id: Id477bfcfccb4dbf5f03f328273c9487ccc48e054