Rename python-keystoneclient-saml2 to keystoneauth
With the ``keystoneauth`` library to be released soon we want our SAML2
authentication plugins to depend on that library, instead of
``python-keystoneclient``. For that we propose changing repository name to
``keystoneauth-saml2`` so it reflects proper hierarchy between
libraries.

This library has never had a release so will not have an impact on
users.

Change-Id: Iddf0e280c1b3513db0700c65e0f912da93119286