Add Solum to OpenStack Projects List
Solum has achived a level of stability and maturity where all of its
contributors feel it is ready to join OpenStack. Solum's mission is to
make cloud services easier to consume and integrate with your application
development process. Adoption of Solum will lead to more application
developers using OpenStack as a cloud platform.

It is directly built on OpenStack infrastructure services including Heat,
Keystone, Glance, and Swift. Through Heat it also uses Nova, and soon
will also use Magnum.

Solum conforms to each of our four opens:
  Open Source
  -----------
  Apache v2.0 License, no dependencies restrict deployment or redistribution.

  Open Development
  ----------------
  We use Stackforge, including Gerrit for all code reviews with a test driven
  gate. The PTL is available as a liason for cross-project teams. We integrate
  with other projects in OpenStack to prevent overlap in our efforts. Solum
  conforms completely to the OpenStack development style.

  Open Design
  -----------
  Our project direction is discussed openly in pubic forums including at
  multiple design summits and midcycle meetups. We use the openstack-dev
  mailing list to discuss issues.

  Open Community
  --------------
  Solum's PTL is chosen by the community in open elections. We hold weekly IRC
  meetings in official OpenStack meeting channels.

We are OpenStack API interoperable, using keystone middleware for discovery
and authentication. We have an active team of dedicated contributors who are
committed to the success of the project.

Change-Id: Id74222381910f006a5a69f9e3746101bd4a32e12