Add type:library tag
The release team is building some tools for automating common release
processes. Some of these tools will look at the type of the repository
to make choices, including skipping or including the repository or
applying different criteria or process steps. For example, changes in
library projects are not available to server projects until the
libraries are released. We want to provide an easy way to find all
unreleased changes in all libraries to ensure that they are released
in a timely manner, and that requires having a way to identify all of
the library projects and their repositories.

Change-Id: I0fb3e09e7d570acdd4a1ef8342814e824d9617c0