Add type:service tag
The release team is building some tools for automating common release
processes. Some of these tools will look at the type of the repository
to make choices, including skipping or including the repository or
applying different criteria or process steps. For example, most of the
service projects use pre-versioning rather than post-versioning, and
so some of the release tools need to take different steps when
processing a service project.

Change-Id: If01633d5b2762cb09ffd80f212e2dcc72156128d