Puppet: new project puppet-openstack-cookiecutter
During last summit, it was decided during the puppet-openstack design session
that the community will provide an easy way for new project to bootstrap their
puppet module with a 'compliant' design.

A cookiecutter template was created[1], after mentioning it on the list[2] and
during our weekly meeting[3] it is a go from the community that this project
should be a core part of the Puppet scope so everyone can contribute to it.

[1] https://github.com/enovance/puppet-openstack-cookiecutter
[2]
http://lists.openstack.org/pipermail/openstack-dev/2015-May/064608.html
[3]
http://eavesdrop.openstack.org/meetings/puppet_openstack/2015/puppet_openstack.2015-06-02-15.00.ht

Change-Id: Iac2f76a0a5704466849cc24b50f70c64653ff932
Depends-On: I18abf0f35f2787b9252a02a9293a669ab1389319