Add neutron to starter-kit:compute
While nova-network is still supported because of concerns about
migration paths, neutron should be as strongly recommended for any new
deployments as is humanly possible so that we can ultimately get
ourselves out of the mess of having two network stacks.

Since starter-kit:compute is intended to help point a finger at what
people should install when trying things out, we should certainly not
imply that they install a network stack that we plan on deleting in the
future. Neutron in provider-vlan/linuxbridge mode is dead simple in not
requiring openvswitch or SDN, and still gets new users a cloud that is
upwards flexible with the future.

Change-Id: I91bce1bcc51121a8a577cff478fdc91002b689b4