Add Fuel to OpenStack Projects
Fuel is an open source deployment and management tool for OpenStack.
Fuel's mission is to streamline and accelerate the otherwise
time-consuming, often complex, and error-prone process of deploying,
testing and maintaining various configurations of OpenStack at scale.

Fuel has been successfully used to deploy OpenStack in environments
ranging from personal proof-of-concept micro-clouds to production
infrastructures composed of hundreds of nodes running tens of thousands
of instances, using a wide variety of network and storage backends, with
out-of-the-box support of most OpenStack projects. Fuel was used by many
successful competitors in the Rule The Stack competition in Vancouver:
https://01.org/openstack/openstacksummitvancouverbc2015/rule-stack-vancouver-results

Fuel is an enabler technology for integration of OpenStack with other
cloud computing initiatives, such as Ceph, OPNFV, Kubernetes:
https://drive.google.com/file/d/0BxYswyvIiAEZUEp4aWJPYVNjeU0/view
https://wiki.opnfv.org/get_started/open_questions#installer_comparison
http://googlecloudplatform.blogspot.com/2015/02/run-container-based-applications-on-OpenStack-with-Kubernetes.html

According to OpenStack user survey before the Liberty OpenStack Summit
(answers collected in March-April 2015), Fuel is the #3 deployment tool
used to set up OpenStack:
http://superuser.openstack.org/articles/openstack-users-share-how-their-deployments-stack-up

Below is a summary of the state of Fuel project's compliance with the
OpenStack Projects requirements as defined in:
http://governance.openstack.org/reference/new-projects-requirements.html

Alignment with OpenStack Mission: Fuel provides a REST API, a web UI, a
command line interface, and plugin framework that automate and simplify
deployment and operation of cloud infrastructures based on OpenStack.

The OpenStack way ("the 4 opens"):

  Open Source: All Fuel components are licensed under Apache License
  v2.0. Fuel does not have library dependencies that restrict how the
  project may be distributed or deployed.

  Open Community: Fuel core reviewers are approved by contributors of
  the project. Fuel project holds weekly IRC meetings on OpenStack
  channels with public meeting agenda, minutes, and logs, since
  February 2014:
  http://eavesdrop.openstack.org/meetings/fuel/

  Open Development: Fuel uses public code reviews on OpenStack
  infrastructure since September 2013:
  https://review.openstack.org/46666

  Commits to all Fuel repositories are approved by core reviewers and
  validated with automated tests on publicly accessible CI:
  https://ci.fuel-infra.org/

  Fuel project has appointed Andrew Woodward as a Community Ambassador
  to serve as a liaison with other OpenStack projects.

  Fuel developers actively engage other OpenStack projects for
  collaboration on topics ranging from high availability to storage:
  http://stackalytics.com/report/contribution/ha-guide/90
  http://lists.openstack.org/pipermail/openstack-dev/2014-July/040484.html

  Fuel project is implemented primarily in Python, adjusts its Python
  dependencies for compliance with OpenStack global requirements, and
  works on reconciling its differences with OpenStack Puppet modules:
  https://blueprints.launchpad.net/fuel/+spec/automate-the-verification-of-compliance-openstack-global-requirement
  https://lwn.net/Articles/648331/

  Open Design: Fuel uses openstack-dev mailing list with [Fuel] tag for
  development discussions. All feature development for Fuel is managed
  via blueprints in Launchpad and specifications in the fuel-specs
  repository on OpenStack infrastructure:
  https://wiki.openstack.org/wiki/Fuel/How_to_contribute#Subscribe_to_mailing_lists
  https://blueprints.launchpad.net/fuel/
  https://review.openstack.org/#/q/project:stackforge/fuel-specs,n,z

Interoperability with the rest of OpenStack: Fuel REST API and its CLI
and web frontends use Keystone for authentication. Fuel REST API has an
overlap with Tuskar, Fuel team would be happy to work with the Tuskar
community on a long-term alignment of APIs if the Technical Committee
determined such a step would be preferable.

Active team: As of July 7, 2015, Fuel project has 23 core reviewers
posting 54.5 reviews per day, and over 200 active contributors:
http://stackalytics.com/report/contribution/fuel-group/90

Change-Id: I8c6579f84a42324eaacae52e459b51c4e7b92b56