Add and apply vulnerability:managed tag
The VMT is building out automation and reporting for vulnerability
management processes in order to better accomodate the rapid growth
of the OpenStack ecosystem. In an order to scale consumability of
its processes beyond its current charter and capacity, a formal
acknowledgement of the list of source code repositories directly
handled by the VMT (rather than managed independently by individual
project teams) is best maintained through application of a
governance-related tag.

Change-Id: I1260839b50b6a5ff57a6649e419094f2d4f7f877