Add the "OpenStack UX" Project:
The full project proposal can be found at
https://wiki.openstack.org/wiki/UX/ProgramProposal

The mission of the UX Program is to support and
facilitate cross-project efforts to improve the overall
user experience of OpenStack.

We provide user research to help teams identify any issues
preventing adoption of their services as well as help the
projects validate design and development efforts to address
those issues. In addition, we help the project teams create
solutions to address customer needs and pain points.

Finally, we will provide the OpenStack community with
visibility into any user experience issues related to
inconsistency across projects.

The UX team is not prescriptive; our goal is to collaborate
with OpenStack’s projects to create better experiences.

Contribution
Member contribution to the UX Project would be measured on
three metrics:

* No. of mock submissions to Invision
* No. of reviews/comments in Invision
* No. of UX studies conducted on behalf of the community

Invision is being provided to OpenStack at no charge as long as
the projects are shared within the open source community.Invision’s
concern was that folks might use our platform for internal
proprietary projects, which is not allowed.

We have 12 administers and 104 managers in Invision that
can add users and create projects from companies including
IBM, Mirantis,  HP, and Redhat.  We were very careful to
make sure that no single person or company had complete
control of the platform.

We are constantly looking for open source alternatives including
phabricator (phabricator.org), which is being discussed for
adoption by the community as a code review tool. Phabricator
also includes a design review that enables users to leave comments
on an image uploaded to the system. An added benefit is that
both code and UX review would be integrated into the same tool.

It is also important to note that projects can be downloaded from
Invision as PDF files that include both images and any comments
from users.  If the community moves away from Invision, the plan
would be to make the PDFs available on the OpenStack UX wiki.

Project Communication
UX meetings were scheduled to occur every other week on IRC
but eventually slowed-down because most of the design efforts
were specifically focused on improving Horizon. That will
eventually change as the UX team engages additional projects.
In addition, IRC is a bit difficult to use within a discipline that is
largely aesthetic.

An additional way to get quick feedback or have a discussion with
the UX group is to join the IRC channel on freenode
(#openstack-ux).

There is a design review that occurs every other week were
community members are invited to share their design work
with the community. These meetings occur via virtual room
rather than IRC because the reviews require the participants to
“see” the designs in order to provide feedback.

A user panel was formed about two months ago in order to
provide feedback to the community from actual users. In
some ways, the user panel is intended to provide an alternative
“voice” to the development community.

The panel is struggling somewhat because we haven’t agreed
on a format for the meetings. For example, one suggestion was
to have a panel members talk about their specific challenges
during a session.

Panel members include users from NSA, Cisco, Pacific Northwest
National Labs, Orange and Yahoo.

Interim PTL
Until we have elections, Pieter Kruithof will be the PTL for delux.

Piet is currently a Sr UX Architect with HP Helion Cloud and
specifically focuses on improving the user experience of OpenStack.
This includes tactical activities such as proposing new designs in
addition to more strategic efforts such as providing platforms that
help enable collaboration within the community.

Piet was also a former Director with the Board of Certification in
Professional Ergonomics (BCPE). The board was established in
1990 as an independent nonprofit organization and is the certifying
body for individuals whose education and experience indicate broad
expertise in the practice of human factors, ergonomics and
user experience research.

Team members include engineers from IBM, Cisco, RackSpace,
HP and Mirantis.

Links
UX wiki
https://wiki.openstack.org/wiki/UX#Horizon_Proposals

User Research wiki
https://wiki.openstack.org/wiki/HorizonUsability_Testing

OpenStack UX patterns
https://wiki.openstack.org/wiki/UX/PatternsLibrary

OpenStack Invision Community
https://openstack.invisionapp.com

Change-Id: I4672d15a8e91190f05ba59086cc3b15f114d5f04